# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_PersonOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PersonOut]",
        "type": "object"
      },
      "PersonIn": {
        "description": "Data of a person; used for create and for full update (PUT).",
        "properties": {
          "birthday": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Birthday as a plain date (YYYY-MM-DD), without time or time zone.",
            "examples": [
              "1994-03-17"
            ],
            "title": "Birthday"
          },
          "name": {
            "description": "Name, 1 to 100 characters.",
            "examples": [
              "Lena"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-text notes, up to 2000 characters. Empty means not set.",
            "examples": [
              "Mag Bouldern und Kaffee, Kleidergr\u00f6\u00dfe M"
            ],
            "title": "Notes"
          },
          "relationship": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relationship to the person, up to 100 characters. Empty means not set.",
            "examples": [
              "Schwester"
            ],
            "title": "Relationship"
          }
        },
        "required": [
          "name"
        ],
        "title": "PersonIn",
        "type": "object"
      },
      "PersonOut": {
        "description": "A person as the frontend sees it.",
        "properties": {
          "birthday": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Birthday as a plain date, or null.",
            "examples": [
              "1994-03-17"
            ],
            "title": "Birthday"
          },
          "created_at": {
            "description": "Creation time (UTC).",
            "examples": [
              "2026-10-04T09:30:00Z"
            ],
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Person ID (UUIDv7).",
            "examples": [
              "0199a8c2-5e3b-7f10-8a4d-2c6e9b1f3a70"
            ],
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "Name.",
            "examples": [
              "Lena"
            ],
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-text notes, or null.",
            "examples": [
              "Mag Bouldern und Kaffee, Kleidergr\u00f6\u00dfe M"
            ],
            "title": "Notes"
          },
          "relationship": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relationship to the person, or null.",
            "examples": [
              "Schwester"
            ],
            "title": "Relationship"
          }
        },
        "required": [
          "id",
          "name",
          "birthday",
          "relationship",
          "notes",
          "created_at"
        ],
        "title": "PersonOut",
        "type": "object"
      },
      "Problem": {
        "description": "Error body for every status outside 2xx (RFC 9457).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "description": "One invalid field of a request.",
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Geschenke-Manager API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/people": {
      "get": {
        "description": "Returns the people of the current account, sorted by name (case-insensitive), one page at a time. `total` counts all people of the account.",
        "operationId": "listPeople",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PersonOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List people",
        "tags": [
          "people"
        ]
      },
      "post": {
        "description": "Creates a person for the current account. Only the name is required; empty optional fields are stored as null. The birthday occasion of the person (type Geburtstag, yearly) follows `birthday`: it is created, updated or deleted with it.",
        "operationId": "createPerson",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create person",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/people/{person_id}": {
      "delete": {
        "description": "Deletes the person and their birthday occasion. Cannot be undone. A person of another account returns 404, like a missing one.",
        "operationId": "deletePerson",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete person",
        "tags": [
          "people"
        ]
      },
      "get": {
        "description": "Returns one person of the current account. A person of another account returns 404, like a missing one.",
        "operationId": "getPerson",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show person",
        "tags": [
          "people"
        ]
      },
      "put": {
        "description": "Replaces all fields of the person: a field left out is reset to null. The frontend sends the whole form. A person of another account returns 404, like a missing one. The birthday occasion of the person (type Geburtstag, yearly) follows `birthday`: it is created, updated or deleted with it.",
        "operationId": "updatePerson",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update person",
        "tags": [
          "people"
        ]
      }
    }
  },
...
}